Refactor monetary values to use float instead of decimal across various models and specs
- Updated specs in `apply_coupon_booking_spec.rb`, `cancel_ticket_booking_spec.rb`, `confirm_payment_booking_spec.rb`, `coupons_operation_spec.rb`, `create_booking_spec.rb`, `create_show_spec.rb` to reflect changes in monetary value handling.
- Introduced new `DateFieldPanel` component for date and time selection in the client.
- Added `VendorShowPerformance` page to display vendor show statistics.
- Created utility functions for date input formatting in `dateInput.js`.
- Implemented job and worker for expiring finished show seat states.
- Added income and shows summary serializers for vendor operations.

+++ server/app/models/booking.rb
@@ -4,19 +4,24 @@ class Booking < ApplicationRecord
   belongs_to :coupon, optional: true
 
   has_many :tickets, dependent: :destroy
+
   has_many :payments, dependent: :destroy
+
   has_many :user_coupon_usages, dependent: :destroy
 
-  enum :status, { pending: "pending", confirmed: "confirmed", cancelled: "cancelled", expired: "expired" }, prefix: true
+  enum :status, { pending: "pending", confirmed: "confirmed",
+    cancelled: "cancelled", expired: "expired" }, prefix: true
 
   before_validation :assign_booking_time, on: :create
 
   validates :booking_time, :status, presence: true
+
   validates :total_amount, numericality: { greater_than_or_equal_to: 0 }
+
   validates :lock_token, presence: true, uniqueness: true
 
   def valid_tickets_subtotal
-    tickets.where(status: "valid").sum(:price).to_d
+    tickets.where(status: "valid").sum(:price).to_f
   end
 
   def refresh_expiration!
